rename notArchived probe to archived
with the swap, the true outcome is now the bad outcome.
