+++ probes/archived/def.yml
@@ -12,17 +12,17 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-id: notArchived
+id: archived
 short: Check that the project is archived
 motivation: >
   A project which is not active might not be patched, have its dependencies patched, or be actively tested and used. However, a lack of active maintenance is not necessarily always a problem. Some software, especially smaller utility functions, does not normally need to be maintained. For example, a library that determines if an integer is even would not normally need maintenance unless an underlying implementation language definition changed. A lack of active maintenance should signal that potential users should investigate further to judge the situation.
 implementation: >
   The probe checks the Archived Status of a project.
 outcome:
-  - If the project is archived, the outcome is OutcomeFalse.
-  - If the project is not archived, the outcome is OutcomeTrue.
+  - If the project is archived, the outcome is OutcomeTrue.
+  - If the project is not archived, the outcome is OutcomeFalse.
 remediation:
-  onOutcome: False
+  onOutcome: True
   effort: High
   text:
     - Non-collaborators, members or owners cannot affect the outcome of this probe.
