+++ probes/archived/impl_test.go
@@ -12,8 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//nolint:stylecheck
-package notArchived
+package archived
 
 import (
 	"testing"
@@ -45,7 +44,7 @@ func Test_Run(t *testing.T) {
 				},
 			},
 			outcomes: []finding.Outcome{
-				finding.OutcomeFalse,
+				finding.OutcomeTrue,
 			},
 		},
 		{
@@ -58,7 +57,7 @@ func Test_Run(t *testing.T) {
 				},
 			},
 			outcomes: []finding.Outcome{
-				finding.OutcomeTrue,
+				finding.OutcomeFalse,
 			},
 		},
 	}
